Extract out the "string interpolation" method from facets
Since I was only using the facets gem just for a single method, it seemed more efficent just to extract out that singular method into its own file, and load that file instead. I provided extensive comments to give credits to the facets gem for their efforts in producing the 'string interpolation' method. I just only wanted that method.

'facets' is a gem that is dual-licensed: both the Ruby license and the GPL applies. I have chosen to follow the Ruby License, as following the GPL license would cause Skynet itself to be a GPL-licesned project (instead of the MIT-license that it currently is), due to the "viral" nature of GPL.